requirements.txt
pandas
tqdm
beautifulsoup4
requests
numpy
python-dateutil
pytest
pdfminer.six
